crawler/flask/requirements.txt
requests
beautifulsoup4
python-dotenv
joblib
